Add orgId to IntegrationProvidier
Integration providers are added with different scopes. Currently it is
cumbersome to query for org scope since only the `teamId` is part of the
integration. This migration would add the `orgId` for org scope and add
constraints that the correct id is set depending on scope.

TODO: The queries accessing the provider need to be adapted to the new
structure.